Expand "uint" to "unsigned int".
"uint" is not a standard type and is not portable.

Also change a few types to "size_t", where appropriate.

Bug: 110161494

+++ hwcrypto/hwkey_srv_fake_provider.c
@@ -159,7 +159,8 @@ static uint32_t get_rpmb_ss_auth_key(const struct hwkey_keyslot* slot,
     if (!rc)
         goto evp_err;
 
-    uint min_kbuf_len = RPMB_SS_AUTH_KEY_SIZE + EVP_CIPHER_CTX_key_length(&evp);
+    size_t min_kbuf_len =
+            RPMB_SS_AUTH_KEY_SIZE + EVP_CIPHER_CTX_key_length(&evp);
     if (kbuf_len < min_kbuf_len) {
         TLOGE("buffer too small: (%zd vs. %zd )\n", kbuf_len, min_kbuf_len);
         goto other_err;
